Learned differnt type of error printing
Learned to differnt methods to show/printing error and different type of
errors.
It is fun to learn and the accomplishment is greate.
Each method brings with its own perks and dis advantges some are
drastic but has good use depending on if you want dicriptive or short
error messages.